requirements.txt
python-dotenv
reedsolo
opencv-python
hmap
numpy
watchfiles
msgpack
# netfilterqueue